requirements.txt
ruamel.yaml==0.17.20
tosca-parser==2.5.0
